+++ nds-model/src/main/java/br/com/abril/nds/dto/ChamadaAntecipadaEncalheDTO.java
@@ -2,6 +2,7 @@ package br.com.abril.nds.dto;
 
 import java.io.Serializable;
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
 public class ChamadaAntecipadaEncalheDTO implements Serializable {
 
@@ -18,7 +19,7 @@ public class ChamadaAntecipadaEncalheDTO implements Serializable {
 	
 	private String nomeCota;
 	
-	private BigDecimal qntExemplares;
+	private BigInteger qntExemplares;
 	
 	private Long codigoChamadaEncalhe;
 	
@@ -26,7 +27,7 @@ public class ChamadaAntecipadaEncalheDTO implements Serializable {
 
 	public ChamadaAntecipadaEncalheDTO() {}
 
-	public ChamadaAntecipadaEncalheDTO(Integer codBox,String nomeBox ,Integer numeroCota,BigDecimal qntExemplares,Long idLancamento,String nomeCota) {
+	public ChamadaAntecipadaEncalheDTO(Integer codBox,String nomeBox ,Integer numeroCota,BigInteger qntExemplares,Long idLancamento,String nomeCota) {
 		
 		this.codBox = codBox;
 		this.numeroCota = numeroCota;
@@ -36,7 +37,7 @@ public class ChamadaAntecipadaEncalheDTO implements Serializable {
 		this.idLancamento = idLancamento;
 	}
 	
-	public ChamadaAntecipadaEncalheDTO(Integer codBox,String nomeBox, Integer numeroCota,BigDecimal qntExemplares,String nomeCota, Long codigoChamadaEncalhe) {
+	public ChamadaAntecipadaEncalheDTO(Integer codBox,String nomeBox, Integer numeroCota,BigInteger qntExemplares,String nomeCota, Long codigoChamadaEncalhe) {
 		
 		this.codBox = codBox;
 		this.numeroCota = numeroCota;
@@ -46,7 +47,7 @@ public class ChamadaAntecipadaEncalheDTO implements Serializable {
 		this.nomeBox = nomeBox;
 	}
 	
-	public ChamadaAntecipadaEncalheDTO(Integer numeroCota,BigDecimal qntExemplares,Long codigoChamadaEncalhe, Long idLancamento) {
+	public ChamadaAntecipadaEncalheDTO(Integer numeroCota,BigInteger qntExemplares,Long codigoChamadaEncalhe, Long idLancamento) {
 		
 		this.numeroCota = numeroCota;
 		this.qntExemplares = qntExemplares;
@@ -143,14 +144,14 @@ public class ChamadaAntecipadaEncalheDTO implements Serializable {
 	/**
 	 * @return the qntExemplares
 	 */
-	public BigDecimal getQntExemplares() {
+	public BigInteger getQntExemplares() {
 		return qntExemplares;
 	}
 
 	/**
 	 * @param qntExemplares the qntExemplares to set
 	 */
-	public void setQntExemplares(BigDecimal qntExemplares) {
+	public void setQntExemplares(BigInteger qntExemplares) {
 		this.qntExemplares = qntExemplares;
 	}
 }
