+++ nds-model/src/main/java/br/com/abril/nds/model/fiscal/ItemNotaFiscalSaida.java
@@ -1,5 +1,6 @@
 package br.com.abril.nds.model.fiscal;
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -53,7 +54,7 @@ public class ItemNotaFiscalSaida {
 	private BigDecimal valorIPIProduto;	
 	
 	@Column(name = "QTDE", nullable = false)
-	private BigDecimal qtde;
+	private BigInteger qtde;
 
 	@ManyToOne(optional = false)
 	@JoinColumn(name = "NOTA_FISCAL_ID")
@@ -71,11 +72,11 @@ public class ItemNotaFiscalSaida {
 		this.id = id;
 	}
 	
-	public BigDecimal getQtde() {
+	public BigInteger getQtde() {
 		return qtde;
 	}
 
-	public void setQtde(BigDecimal qtde) {
+	public void setQtde(BigInteger qtde) {
 		this.qtde = qtde;
 	}
